requirements.txt
python-dotenv
flask
requests